services/lead_router/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
aioredis==2.0.1
pyyaml==6.0
httpx==0.27.0